+++ app/components/Views/SendFlow/AddressFrom/AddressFrom.test.tsx
@@ -1,11 +1,12 @@
-import React from 'react';
-import configureMockStore from 'redux-mock-store';
 import { render } from 'enzyme';
-import SendToAddressTo from './';
+import React from 'react';
 import { Provider } from 'react-redux';
-import Engine from '../../../../core/Engine';
+import configureMockStore from 'redux-mock-store';
 
-Engine.init();
+import Engine from '../../../../core/Engine';
+import SendToAddressFrom from './';
+
+Engine.init({});
 jest.mock('@react-navigation/native', () => ({
   useNavigation: () => ({
     navigation: {},
@@ -24,9 +25,6 @@ const initialState = {
           },
         },
       },
-      AddressBookController: {
-        addressBook: {},
-      },
       PreferencesController: {
         selectedAddress: '0x0',
         identities: {
@@ -40,9 +38,6 @@ const initialState = {
   },
 };
 
-const mockStore = configureMockStore();
-const store = mockStore(initialState);
-
 jest.mock('react-redux', () => ({
   ...jest.requireActual('react-redux'),
   useSelector: jest
@@ -50,23 +45,14 @@ jest.mock('react-redux', () => ({
     .mockImplementation((callback) => callback(initialState)),
 }));
 
-describe('SendToAddressTo', () => {
+const mockStore = configureMockStore();
+const store = mockStore(initialState);
+
+describe('SendToAddressFrom', () => {
   it('should render correctly', () => {
     const wrapper = render(
       <Provider store={store}>
-        <SendToAddressTo
-          inputRef={undefined}
-          highlighted={false}
-          addressToReady={false}
-          toSelectedAddress={undefined}
-          toSelectedAddressName={undefined}
-          onSubmit={() => undefined}
-          inputWidth={undefined}
-          confusableCollectionArray={undefined}
-          isFromAddressBook={undefined}
-          updateParentState={undefined}
-          onToSelectedAddressChange={undefined}
-        />
+        <SendToAddressFrom fromAccountBalanceState={() => undefined} />
       </Provider>,
     );
     expect(wrapper).toMatchSnapshot();
