+++ app/components/Views/SendFlow/AddressFrom/index.ts
@@ -1,1 +1,1 @@
-export { default } from './SendToAddressFrom';
+export { default } from './AddressFrom';
