+++ app/components/Views/SendFlow/AddressTo/AddressTo.types.ts
@@ -1,0 +1,13 @@
+export interface STAddressToProps {
+  addressToReady: boolean;
+  confusableCollectionArray: any;
+  highlighted: boolean;
+  inputRef: any;
+  inputWidth: any;
+  isFromAddressBook: any;
+  onSubmit: (address: string) => void;
+  onToSelectedAddressChange: any;
+  toSelectedAddress: any;
+  toSelectedAddressName: any;
+  updateParentState: any;
+}
