--- app/components/Views/SendFlow/AddressTo/SendToAddressTo.types.ts
@@ -1,13 +1,0 @@
-export interface STAddressToProps {
-  inputRef: any;
-  highlighted: boolean;
-  addressToReady: boolean;
-  toSelectedAddress: any;
-  toSelectedAddressName: any;
-  onSubmit: (address: string) => void;
-  inputWidth: any;
-  confusableCollectionArray: any;
-  isFromAddressBook: any;
-  updateParentState: any;
-  onToSelectedAddressChange: any;
-}
