+++ app/components/Views/SendFlow/AddressTo/index.ts
@@ -1,1 +1,1 @@
-export { default } from './SendToAddressTo';
+export { default } from './AddressTo';
